Error Querying when "+" found in title
The query would return an empty list when there is a "+" in the query. With a little experimentation, I found that if you replace it with a _ it works. 
Ex
https://myanimelist.net/api/anime/search.xml?q=A-Channel:+_A-Channel
and 
https://myanimelist.net/api/anime/search.xml?q=A-Channel:++A-Channel
As you can see, the first query correctly returns information on the anime, while the second one returns nothing.